Consolidated Course Structure API Client
This consolidations removes duplicated code and adds the ability for the access token to be overridden by Django settings. The override should make it easier to specify an access token if the authenticated user does not have one, specifically when running tests locally using a course structure API client on another machine.