WIP created new common generic-list app.
Moved these files to generic-list (and removed them from learners and
course-list) so far:

* header.js, header.underscore
* alert-view.js, alert.underscore
* page.js, page-spec.js
* download-data.js, download-data-spec.js, download-data.underscore

Also, updated the require() calls in learners and course-list to point to
generic-list for the files above.

Still need to move many more files to generic-list and update their references
in learners and course-list.

I tried and failed to move the app.js for each app to generic-list. We will have
to keep those in separate apps.