eucalyptus: Pin path.py for old i18n-tools version
The version of i18n-tools that we are using on this branch does not pin its
path.py requirement and it was pulling a version that was too new. We fixed
this in master by upgrading i18n-tools, but that involves a larger change that
should go in the next OpenEdx release.