Update NewRelic in Insights
### [EDUCATOR-3734](https://openedx.atlassian.net/browse/EDUCATOR-3734)
Update NewRelic on Insights.

As `newrelic` was pinned to a specific version, I removed the pin and run `make upgrade` to update it to the latest version of newrelic.
I also needed to make a additional change in `Makefile` to move `test.in` above `local.in` as `local.in` needs `test.txt`.