feat: switch most enrollments views to v1
Enrollments views are now all switched to v1 with the exception of age
which is not supported in the v1 API currently. This required pushing
the v0/v1 choice down a level in the demographics inheritance tree.
The course list is switched over as well.

Also switches CSVs including one missed for activities.

MST-1250